requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
bandit==1.6.2
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.3
coverage==5.0.3
docutils==0.16
dparse==0.4.1
gitdb2==3.0.0
GitPython==3.0.7
idna==2.8
isort==4.3.21
keyring==21.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.1
pbr==5.4.4
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
Pygments==2.5.2
pyinstrument==3.1.0
pyinstrument-cext==0.2.2
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
safety==1.8.5
six==1.14.0
smmap2==2.0.5
snowballstemmer==2.0.0
stevedore==1.32.0
tqdm==4.42.1
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
